requirements.txt
diffusers
transformers
accelerate
torchvision
torchaudio
sentencepiece